requirements.txt
Tornado==3.0.2
Mongoengine
unidecode
py-bcrypt
numpy
python-geoip
python-geoip-geolite2